% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-solution.R
\name{aggregate_survey_response_by_county}
\alias{aggregate_survey_response_by_county}
\title{Aggregate survey responses by county}
\usage{
aggregate_survey_response_by_county(ling_data)
}
\arguments{
\item{ling_data}{the linguistics data}
}
\value{
A data frame with one row per county and each column is a question.
The (i, j) value in this data frame is the most popular response to
question j in county i.
}
\description{
Aggregate survey responses by county
}
