% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-solution.R
\name{clean_ling_data}
\alias{clean_ling_data}
\title{Clean linguistics data}
\usage{
clean_ling_data(ling_data, qa_key = NULL)
}
\arguments{
\item{ling_data}{the merged linguistics data}
}
\value{
A cleaned data frame
}
\description{
Clean linguistics data
}
