% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_clustering_utils.R
\name{clusteringOptionsUI}
\alias{clusteringOptionsUI}
\title{Widgets to set data and clustering options}
\usage{
clusteringOptionsUI(id)
}
\description{
Widgets to set data and clustering options
}
\keyword{Dimension}
\keyword{internal}
\keyword{options}
\keyword{reduction}
