% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-solution.R
\name{collapse_survey_responses}
\alias{collapse_survey_responses}
\title{Collapse rare survey responses into an "other" category}
\usage{
collapse_survey_responses(ling_data, qa_key, min_prop = 0.05)
}
\arguments{
\item{ling_data}{the linguistics data}

\item{qa_key}{the question-answer key}

\item{min_prop}{collapse responses with a proportion less than this value
into an "other" category}
}
\value{
A list of two: (1) the collapsed data frame and (2) the updated
question-answer key
}
\description{
Collapse rare survey responses into an "other" category
}
