% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_answers}
\alias{get_answers}
\title{Convert numeric answers to text}
\usage{
get_answers(x, qid, qa_key)
}
\arguments{
\item{x}{the numeric answers}

\item{qid}{the question ID}

\item{qa_key}{the question-answer key}
}
\value{
a factor with text answers
}
\description{
Convert numeric answers to text
}
