% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_X_matrix}
\alias{get_X_matrix}
\title{Get a numeric X matrix from the linguistic data}
\usage{
get_X_matrix(ling_data)
}
\arguments{
\item{ling_data}{the linguistic data}
}
\value{
a numeric matrix
}
\description{
Get a numeric X matrix from the linguistic data
}
