% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-solution.R
\name{keep_contiguous_us}
\alias{keep_contiguous_us}
\title{Remove samples from Hawaii and Alaska}
\usage{
keep_contiguous_us(ling_data)
}
\arguments{
\item{ling_data}{the linguistics data}
}
\value{
A data frame with samples from Hawaii and Alaska removed
}
\description{
Remove samples from Hawaii and Alaska
}
