% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-solution.R
\name{load_ling_data}
\alias{load_ling_data}
\title{Load in lingusitics data}
\usage{
load_ling_data()
}
\arguments{
\item{path}{the path to the data}
}
\value{
A data frame containing the linguistics survey data
}
\description{
Load in lingusitics data
}
