% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-solution.R
\name{load_q_and_a_key}
\alias{load_q_and_a_key}
\title{Load in question and answer key}
\usage{
load_q_and_a_key()
}
\arguments{
\item{path}{the path to the data}
}
\value{
A data frame containing the question and answer key
}
\description{
Load in question and answer key
}
