% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-solution.R
\name{load_zip_data}
\alias{load_zip_data}
\title{Load in ZIP locations data}
\usage{
load_zip_data(path = here::here("data"))
}
\arguments{
\item{path}{the path to the data}
}
\value{
A data frame containing the ZIP locations data
}
\description{
Load in ZIP locations data
}
