% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-solution.R
\name{merge_ling_zip_data}
\alias{merge_ling_zip_data}
\title{Merge linguistics data with ZIP location data}
\usage{
merge_ling_zip_data(ling_data, zip_data)
}
\arguments{
\item{ling_data}{the linguistics data}

\item{zip_data}{the ZIP data}
}
\value{
A merged data frame
}
\description{
Merge linguistics data with ZIP location data
}
