% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-solution.R
\name{one_hot_ling_data}
\alias{one_hot_ling_data}
\title{One-hot encoding of categorical variables}
\usage{
one_hot_ling_data(ling_data, remove_zeros = FALSE, return_matrix = TRUE)
}
\arguments{
\item{ling_data}{the linguistics data}

\item{remove_zeros}{whether to remove .0 columns which correspond to NAs}

\item{return_matrix}{whether to return a matrix including only the survey
responses or the full data frame (including location information)}
}
\value{
A one-hot-encoded data matrix
}
\description{
One-hot encoding of categorical variables
}
