% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dr_map}
\alias{plot_dr_map}
\title{Plot dimension reduction results on a map}
\usage{
plot_dr_map(
  X,
  ling_data,
  ndim = 2,
  point_size = 0.5,
  point_alpha = 0.5,
  linewidth = 0.1,
  by_county = FALSE,
  show_contiguous_us = TRUE
)
}
\arguments{
\item{X}{the data matrix}

\item{ling_data}{the linguistic data}

\item{ndim}{the number of top components to plot}

\item{point_size}{the size of the points}

\item{point_alpha}{the transparency of the points}

\item{linewidth}{the width of the county borders}

\item{by_county}{whether to plot by county}

\item{show_contiguous_us}{whether to show only the contiguous US}
}
\value{
a ggplot object
}
\description{
Plot dimension reduction results on a map
}
