% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dr_scatter}
\alias{plot_dr_scatter}
\title{Plot dimension reduction results}
\usage{
plot_dr_scatter(
  X,
  ndim = 2,
  color = NULL,
  point_size = 0.5,
  point_alpha = 0.5,
  ...
)
}
\arguments{
\item{X}{the data matrix}

\item{ndim}{the number of top components to plot}

\item{color}{a vector to use for coloring data points}

\item{point_size}{the size of the points}

\item{point_alpha}{the transparency of the points}

\item{...}{additional arguments to pass to vdocs::plot_pairs()}
}
\value{
a ggplot object
}
\description{
Plot dimension reduction results
}
