% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ling_map}
\alias{plot_ling_map}
\title{Plot linguistic survey results on US map}
\usage{
plot_ling_map(
  ling_data,
  qa_key,
  qid,
  point_size = 0.5,
  map_type = c("state", "county"),
  show_contiguous_us = TRUE
)
}
\arguments{
\item{ling_data}{the linguistic survey data}

\item{qa_key}{the question and answer key}

\item{qid}{the question ID}

\item{point_size}{the size of the points on the map}

\item{map_type}{the type of map to plot}

\item{show_contiguous_us}{whether to show only the contiguous US}
}
\value{
a ggplot object
}
\description{
Plot linguistic survey results on US map
}
