% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ling_map_by_county}
\alias{plot_ling_map_by_county}
\title{Plot linguistic survey results, aggregated by county}
\usage{
plot_ling_map_by_county(
  ling_data,
  qa_key,
  qid,
  linewidth = 0.1,
  show_contiguous_us = TRUE
)
}
\arguments{
\item{ling_data}{the linguistic survey data}

\item{qa_key}{the question and answer key}

\item{qid}{the question ID}

\item{linewidth}{the width of the county borders}

\item{show_contiguous_us}{whether to show only the contiguous US}
}
\value{
a ggplot object
}
\description{
Plot linguistic survey results, aggregated by county
}
