% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA results}
\usage{
plot_pca(
  pca_out,
  npcs = 2,
  color = NULL,
  point_size = 0.5,
  point_alpha = 0.5,
  ...
)
}
\arguments{
\item{pca_out}{the PCA output}

\item{npcs}{the number of principal components to plot}

\item{color}{a vector to use for coloring data points}

\item{point_size}{the size of the points}

\item{point_alpha}{the transparency of the points}

\item{...}{additional arguments to pass to vdocs::plot_pairs()}
}
\value{
a ggplot object
}
\description{
Plot PCA results
}
