% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-solution.R
\name{remove_samples}
\alias{remove_samples}
\title{Remove samples with too many missing values}
\usage{
remove_samples(ling_data, min_answers = 50)
}
\arguments{
\item{ling_data}{the linguistics data}

\item{min_answers}{the minimum number of answers required for a sample
to be retained}
}
\value{
A data frame with samples removed
}
\description{
Remove samples with too many missing values
}
