% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_clustering.R
\name{run_app}
\alias{run_app}
\title{Run hierarchical clustering Shiny application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Additional arguments to pass to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\value{
A shiny application
}
\description{
Run hierarchical clustering Shiny application
}
