% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_clustering_utils.R
\name{updateClusteringOptions}
\alias{updateClusteringOptions}
\title{Update clustering option widgets}
\usage{
updateClusteringOptions(id)
}
\description{
Update clustering option widgets
}
\keyword{internal}
